Refactor: RecordDetail.vue -> 날짜 탭 UI 기능분리
